Django-REST/requirements.txt
Django==1.10.2
djangorestframework==3.5.1
psycopg2==2.6.2
django-nose==1.4
coverage==3.7.1
